requirements.txt
anthropic>=0.39.0
psycopg2-binary>=2.9.0
python-pptx>=1.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit>=1.32
mpxj>=15.0.0
jpype1>=1.5.0
pytest>=7.0
pytest-cov>=4.0
ruff>=0.6.0
